feat: Move Data Classes to commonMain
