+++ core/model/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/model/data/FollowableTopic.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,13 +17,10 @@
 package com.google.samples.apps.nowinandroid.core.model.data
 
 /**
- * External data layer representation of a NiA Topic
+ * A [topic] with the additional information for whether or not it is followed.
  */
-data class Topic(
-    val id: String,
-    val name: String,
-    val shortDescription: String,
-    val longDescription: String,
-    val url: String,
-    val imageUrl: String,
+// TODO consider changing to UserTopic and flattening
+data class FollowableTopic(
+    val topic: Topic,
+    val isFollowed: Boolean,
 )
