+++ core/model/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/model/data/SearchResult.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,8 @@
 
 package com.google.samples.apps.nowinandroid.core.model.data
 
-/**
- * An entity of [SearchResult] with additional user information such as whether the user is
- * following a topic.
- */
-data class UserSearchResult(
-    val topics: List<FollowableTopic> = emptyList(),
-    val newsResources: List<UserNewsResource> = emptyList(),
+/** An entity that holds the search result */
+data class SearchResult(
+    val topics: List<Topic> = emptyList(),
+    val newsResources: List<NewsResource> = emptyList(),
 )
