+++ core/model/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/model/data/UserData.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,15 @@
 
 package com.google.samples.apps.nowinandroid.core.model.data
 
-enum class DarkThemeConfig {
-    FOLLOW_SYSTEM,
-    LIGHT,
-    DARK,
-}
+/**
+ * Class summarizing user interest data
+ */
+data class UserData(
+    val bookmarkedNewsResources: Set<String>,
+    val viewedNewsResources: Set<String>,
+    val followedTopics: Set<String>,
+    val themeBrand: ThemeBrand,
+    val darkThemeConfig: DarkThemeConfig,
+    val useDynamicColor: Boolean,
+    val shouldHideOnboarding: Boolean,
+)
