+++ core/model/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/model/data/UserNewsResource.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
